requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Pillow==10.0.0
reportlab==4.0.4
qrcode==7.4.2
pandas==2.1.0
numpy==1.25.2
